refactor : SPR 원칙을 위해 커스텀 훅으로 분리
