+++ apps/frontend/app/(route)/profile/_components/ProblemHeatmap.tsx
@@ -1,91 +1,32 @@
-import { useState, useEffect } from "react";
 import HeatMap, { ReactCalendarHeatmapValue } from "react-calendar-heatmap";
 import clsx from "clsx";
 
 import { useAuthStore } from "@/app/_store/authStore";
+import { useProblems } from "@/app/_hooks/useProblems";
 
 import { Problem, HeatmapValue } from "@/app/_type/problem";
 import { COLOR_SCALE } from "@/app/_constants/colorScale";
-import {
-  fetchUserData,
-  fetchProblems,
-  fetchProblemsByDate,
-} from "@/app/api/problem";
 
 interface ProblemHeatmapProps {
   onDateSelect: (date: string, problems: Problem[]) => void;
 }
 
 export default function ProblemHeatmap({ onDateSelect }: ProblemHeatmapProps) {
-  const [data, setData] = useState<HeatmapValue[]>([]);
-  const [years, setYears] = useState<number[]>([]);
-  const [selectedYear, setSelectedYear] = useState<number | null>(null);
-
   const token = useAuthStore((state) => state.token);
-
-  useEffect(() => {
-    if (!token) return;
-
-    const loadUser = async () => {
-      try {
-        const data = await fetchUserData(token);
-        const startYear = new Date(data.createdAt).getFullYear();
-        const currentYear = new Date().getFullYear();
-        const yearList = Array.from(
-          { length: currentYear - startYear + 1 },
-          (_, i) => startYear + i,
-        );
-
-        setYears(yearList);
-        setSelectedYear(currentYear);
-      } catch (error) {
-        console.error("사용자 데이터 불러오기 실패:", error);
-      }
-    };
-
-    loadUser();
-  }, [token]);
-
-  useEffect(() => {
-    if (!token || !selectedYear) return;
-
-    const loadProblems = async () => {
-      try {
-        const result = await fetchProblems(token);
-
-        const mapped: HeatmapValue[] = result.reduce<HeatmapValue[]>(
-          (acc, item) => {
-            const date = item.createdAt.split("T")[0];
-            const existing = acc.find((d) => d.date === date);
-
-            if (existing) {
-              existing.count += 1;
-            } else {
-              acc.push({ date, count: 1 });
-            }
-            return acc;
-          },
-          [],
-        );
-
-        setData(mapped);
-      } catch (err) {
-        console.error("❌ API 호출 에러:", err);
-      }
-    };
-
-    loadProblems();
-  }, [token, selectedYear]);
+  const {
+    years,
+    selectedYear,
+    setSelectedYear,
+    data,
+    loading,
+    error,
+    fetchByDate,
+  } = useProblems(token);
 
   const handleDayClick = async (value?: ReactCalendarHeatmapValue<string>) => {
-    if (!value?.date) return;
-
-    try {
-      const result = await fetchProblemsByDate(token!, value.date);
-      onDateSelect(value.date, result);
-    } catch (err) {
-      console.error("❌ 날짜별 문제 불러오기 실패:", err);
-    }
+    if (!value?.date || !token) return;
+    const result = await fetchByDate(token, value.date);
+    onDateSelect(value.date, result);
   };
 
   const startDate = selectedYear
@@ -95,6 +36,9 @@ export default function ProblemHeatmap({ onDateSelect }: ProblemHeatmapProps) {
 
   return (
     <div className="mt-6 overflow-x-auto p-4 pl-20 shadow-md">
+      {loading && <p className="text-gray-500">데이터 불러오는 중...</p>}
+      {error && <p className="text-red-500">{error}</p>}
+
       <div className="mx-auto flex w-fit">
         <div className="my-auto flex h-full flex-col justify-center space-y-2 p-4 pl-0">
           {years.map((year) => (
